Make pytest fixtures reusable
